First revisions after reviewing diff with master. Removing old Contract contruction requests and cleaning up some duplicated code in the DuesRequirement(s) classes.
