+++ app/ContractFramework/Requirements/AssociateMemberDuesRequirement.php
@@ -1,14 +1,6 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: james
- * Date: 9/24/15
- * Time: 11:54 PM
- */
 
 namespace APOSite\ContractFramework\Requirements;
-
-use APOSite\Models\Semester;
 
 class AssociateMemberDuesRequirement extends DuesBaseRequirement
 {
@@ -17,26 +9,4 @@ class AssociateMemberDuesRequirement extends DuesBaseRequirement
 
     protected $threshold = 35;
     protected $comparison = 'GEQ';
-
-    public function computeValue()
-    {
-        $reports = $this->getReports($this->semester);
-        if ($reports->isEmpty()) {
-            return 0;
-        } else {
-            return $reports->last()->pivot->value;
-        }
-    }
-
-    public function getReports()
-    {
-        $dues_reports = $this->user->reports()->DuesReports()->get();
-        $semester = Semester::currentSemester();
-        $dues_reports = $dues_reports->filter(function ($report) use ($semester) {
-            $val = $semester->dateInSemester($report->report_type->report_date);
-            return $val;
-        });
-        return $dues_reports;
-    }
-
 }
