+++ app/ContractFramework/Requirements/DuesBaseRequirement.php
@@ -14,4 +14,25 @@ abstract class DuesBaseRequirement extends Requirement
     {
         return view('reports.duesreports.dueseventlist');
     }
+
+    public function computeValue()
+    {
+        $reports = $this->getReports($this->semester);
+        if ($reports->isEmpty()) {
+            return 0;
+        } else {
+            return $reports->last()->pivot->value;
+        }
+    }
+
+    public function getReports()
+    {
+        $dues_reports = $this->user->reports()->DuesReports()->get();
+        $semester = Semester::currentSemester();
+        $dues_reports = $dues_reports->filter(function ($report) use ($semester) {
+            $val = $semester->dateInSemester($report->report_type->report_date);
+            return $val;
+        });
+        return $dues_reports;
+    }
 }
